Reset: Remove standard Google Map and Places API scripts
Although the standard script provided an easy win as far as generating functionality, it quickly became a bit too much to manage as it was hard to determine where all the dependencies were coming from. As a result, I have removed it and plan to build it incrementally for a better code base.